Add files via upload
The S.T.I.T.C.H. demo is a foundation for a garment designing tool. S.T.I.T.C.H. demo allows the user to create their own project, currently limited to making boxer shorts as they are the simplest garment to make for anyone starting out in sewing. After providing a base with measurements as a size reference, the user can fill the panel with any desired design from an assortment of colors to their own choice of designs in the form of an image. The user also has many save options being able to name their project at will, save their project into an assembly of organized folders, and retrieve any previous saved project to display and even edit. Attached in this submission are some example files for customizable filling designs, but you can use any design that you wish.

The project requires a 'Projects' folder and a 'Designs' folder in the same relative path of the program. While the program will make a 'Projects' folder in the event that it doesn't exist, **calling upon the design feature with no existing 'Design' folder or no contents in the 'Design folder' will render the customizable aspect of the feature UNUSABLE**. Please be aware of this before running the program.